requirements.txt
asgiref==3.5.2
bleach==5.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1.1
django-ckeditor==6.5.0
django-extra-fields==3.0.2
django-js-asset==2.0.0
django-random-filestorage==0.1.0
django-yaturbo==1.0.1
djangorestframework==3.13.1
lxml==4.9.1
Pillow==9.2.0
pytils==0.4.1
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
webencodings==0.5.1
